feat: implementar backend completo de alquiler de vehículos
## Autenticación
- Registro de usuarios por rol (CLIENTE, EMPLEADO, ADMIN)
- Login con JWT y detección de rol por herencia JPA
- JwtUtil con JJWT 0.11.5 (parserBuilder + Keys.hmacShaKeyFor)
- JwtAuthFilter (OncePerRequestFilter) para validar token en cada request
- UserDetailsServiceImpl para cargar usuario desde BD
- AuthUtils para extraer usuario/cliente/empleado del SecurityContext
- SecurityConfig con reglas de acceso por rol y método HTTP

## Modelos / Entidades
- Usuario (herencia JOINED) → Cliente, Empleado, Administrador
- Vehiculo, CategoriaVehiculo, Sucursal
- Reserva, Alquiler, Pago, Multa, Mantenimiento
- Todos los IDs como Long con @Column(name = "id_*")
- Campos con snake_case y @Column(name = ...) para mapeo a BD

## Repositorios
- IUsuarioRepository, IClienteRepository, IEmpleadoRepository, IAdministradorRepository
- IVehiculoRepository, ICategoriaVehiculoRepository, ISucursalRepository
- IReservaRepository, IAlquilerRepository, IPagoRepository
- IMultaRepository, IMantenimientoRepository
- @Query con JPQL para rangos de fechas y cancelaciones automáticas

## DTOs
- Request y Response para cada módulo
- @JsonProperty para campos snake_case
- @NotBlank, @NotNull, @Email para validaciones
- Sin @Builder en DTOs de request para compatibilidad con Jackson

## Mapper
- Clase estática con métodos toEntity() y toDTO()
- Manejo de herencia con switch por instanceof
- Cálculo de totalEstimado con ChronoUnit.DAYS

## Servicios e Interfaces
- IAuthService / AuthService
- IVehiculoService / VehiculoService
- ICategoriaVehiculoService / CategoriaVehiculoService
- ISucursalService / SucursalService
- IReservaService / ReservaService
- IAlquilerService / AlquilerService
- IPagoService / PagoService
- IMultaService / MultaService
- IMantenimientoService / MantenimientoService
- IUsuarioService / UsuarioService

## Flujo de negocio implementado
- Cliente busca vehículos disponibles por rango de fechas
- Cliente crea reserva → vehículo pasa a RESERVADO
- Cliente confirma reserva → se crea Alquiler en PENDIENTE_PAGO
- Cliente paga → Alquiler pasa a ACTIVO automáticamente
- Empleado finaliza alquiler → vehículo vuelve a DISPONIBLE
- Empleado registra multas en alquileres ACTIVOS o FINALIZADOS
- Cancelación automática a medianoche con @Scheduled

## Cancelación automática
- ScheduledService con @Scheduled(cron = "0 0 0 * * *")
- Cancela reservas PENDIENTES cuya fecha_inicio ya llegó
- Cancela alquileres PENDIENTE_PAGO cuya fecha_inicio ya llegó
- @EnableScheduling en clase principal

## Manejo de excepciones
- GlobalExceptionHandler con @RestControllerAdvice
- ErrorResponseDTO con status, mensaje y timestamp
- Manejo de RuntimeException, NoSuchElementException
- Manejo de MethodArgumentNotValidException, AccessDeniedException
- Manejo de ExpiredJwtException

## Controllers
- AuthController
- VehiculoController
- CategoriaVehiculoController
- SucursalController
- ReservaController
- AlquilerController
- PagoController
- MultaController
- MantenimientoController
- UsuarioController

## Seguridad por rol
- GET /api/vehiculos/** → ADMIN, EMPLEADO, CLIENTE
- POST/PUT/DELETE /api/vehiculos/** → ADMIN, EMPLEADO
- /api/reservas/** → ADMIN, EMPLEADO, CLIENTE
- /api/pagos/** → ADMIN, EMPLEADO, CLIENTE
- /api/alquileres/** → ADMIN, EMPLEADO
- /api/mantenimientos/** → ADMIN, EMPLEADO
- /api/usuarios/clientes → ADMIN
- /api/usuarios/empleados → ADMIN
- /api/auth/** → público

+++ Backend/alquiler_carros/src/main/java/base_datos/alquiler_carros/Repository/IMantenimientoRepository.java
@@ -13,10 +13,10 @@ import java.util.List;
 @Repository
 public interface IMantenimientoRepository extends JpaRepository<Mantenimiento, Long> {
 
-    List<Mantenimiento> findByVehiculo_Id(Integer id_vehiculo);
+    List<Mantenimiento> findByVehiculo_Id(Long id_vehiculo);
     List<Mantenimiento> findByFechaBetween(LocalDate fecha_inicio, LocalDate fecha_fin);
 
     // Costo total de mantenimientos de un vehículo
     @Query("SELECT SUM(m.costo) FROM Mantenimiento m WHERE m.vehiculo.id = :id_vehiculo")
-    BigDecimal sumCostoByVehiculo(@Param("id_vehiculo") Integer id_vehiculo);
+    BigDecimal sumCostoByVehiculo(@Param("id_vehiculo") Long id_vehiculo);
 }
